Refactor style checks
