+++ lookout/style/format/code_generator.py
@@ -8,7 +8,8 @@ from lookout.style.format.feature_extractor import FeatureExtractor
 from lookout.style.format.feature_utils import (
     CLASS_INDEX, CLASS_REPRESENTATIONS, CLASSES, CLS_DOUBLE_QUOTE, CLS_NEWLINE, CLS_NOOP,
     CLS_SINGLE_QUOTE, CLS_SPACE_DEC, CLS_SPACE_INC, CLS_TAB_DEC, CLS_TAB_INC, CLS_TO_STR,
-    VirtualNode)
+    VirtualNode
+)
 
 
 class CodeGenerationError(Exception):
