+++ lookout/style/format/descriptions.py
@@ -7,9 +7,8 @@ from typing import List, Optional, Sequence, Tuple
 from numpy import flatnonzero, floating, ndarray
 
 from lookout.style.format.feature_extractor import FeatureExtractor, FEATURES_MAX, FEATURES_MIN
-from lookout.style.format.feature_utils import (
-    CLASS_INDEX, CLASS_PRINTABLES, CLASS_REPRESENTATIONS, CLS_NOOP)
-from lookout.style.format.feature_utils import VirtualNode
+from lookout.style.format.feature_utils import CLASS_INDEX, CLASS_PRINTABLES, \
+    CLASS_REPRESENTATIONS, CLS_NOOP, VirtualNode
 from lookout.style.format.features import BagFeature, CategoricalFeature, OrdinalFeature
 from lookout.style.format.rules import Rule
