+++ lookout/style/format/research/seq_matching_vis.py
@@ -4,8 +4,8 @@ from difflib import SequenceMatcher
 import os
 
 from bblfsh import BblfshClient
-
 from lookout.core.api.service_data_pb2 import File
+
 from lookout.style.format.feature_extractor import FeatureExtractor
 from lookout.style.format.feature_utils import CLASSES
 from lookout.style.format.model import FormatModel
