+++ lookout/style/format/tests/test_analyzer_integration.py
@@ -9,10 +9,11 @@ from threading import Thread
 import unittest
 from unittest.mock import patch
 
-from lookout.__main__ import main as launch_analyzer
 from lookout.core.event_listener import EventListener
-from lookout.core.tests.server import find_port, run as launch_server
+from lookout.core.test_helpers.server import find_port, run as launch_server
 
+# PyCharm has a bug and inserts a new line here
+from lookout.__main__ import main as launch_analyzer  # noqa: I202, I100
 
 FROM_COMMIT = "HEAD" + "^" * 8
 TO_COMMIT = "HEAD"
