+++ lookout/style/format/tests/test_quality_report.py
@@ -7,10 +7,11 @@ import tarfile
 import tempfile
 import unittest
 
-from lookout.core.tests.server import find_port, run as launch_server
+from lookout.core.test_helpers.server import find_port, run as launch_server
+
 from lookout.style.format.quality_report import quality_report
-from lookout.style.format.tests.test_analyzer_integration import (FROM_COMMIT, TestAnalyzer,
-                                                                  TO_COMMIT)
+from lookout.style.format.tests.test_analyzer_integration import (
+    FROM_COMMIT, TestAnalyzer, TO_COMMIT)
 
 
 class PretrainedModelTests(unittest.TestCase):
