+++ lookout/style/format/tests/test_rule_stat.py
@@ -2,10 +2,11 @@ import os
 import tempfile
 import unittest
 
-from lookout.core.tests.server import find_port, run as launch_server
+from lookout.core.test_helpers.server import find_port, run as launch_server
+
 from lookout.style.format.rule_stat import print_rules_report
-from lookout.style.format.tests.test_analyzer_integration import (FROM_COMMIT, TestAnalyzer,
-                                                                  TO_COMMIT)
+from lookout.style.format.tests.test_analyzer_integration import (
+    FROM_COMMIT, TestAnalyzer, TO_COMMIT)
 from lookout.style.format.tests.test_quality_report import Capturing, PretrainedModelTests
