+++ lookout/style/typos/research/create_typos.py
@@ -6,7 +6,6 @@ from tqdm import tqdm
 
 from lookout.style.typos.research.typos_functions import rand_bool
 
-
 letters = list(string.ascii_lowercase)
