+++ lookout/style/typos/research/main.py
@@ -1,13 +1,13 @@
 import argparse
 import sys
 
-from lookout.style.typos.research.create_typos import create_typos
+from sourced.ml.cmd.args import ArgumentDefaultsHelpFormatterNoNone
+
 from lookout.style.typos.research.baseline import baseline
+from lookout.style.typos.research.create_typos import create_typos
 from lookout.style.typos.research.filter_identifiers import filter_identifiers
 from lookout.style.typos.research.get_frequencies import get_frequencies
 from lookout.style.typos.research.pick_subset import pick_subset
-
-from sourced.ml.cmd.args import ArgumentDefaultsHelpFormatterNoNone
 
 
 def get_parser() -> argparse.ArgumentParser:
