+++ lookout/style/typos/tests/test_corrector.py
@@ -7,7 +7,6 @@ import pandas
 from lookout.style.typos.corrector import TyposCorrector
 from lookout.style.typos.utils import CORRECT_TOKEN_COLUMN, SPLIT_COLUMN, TYPO_COLUMN
 
-
 TEST_DATA_PATH = str(pathlib.Path(__file__).parent)
 FASTTEXT_DUMP_FILE = str(pathlib.Path(__file__).parent / "id_vecs_10.bin")
 VOCABULARY_FILE = str(pathlib.Path(__file__).parent / "test_frequencies.csv.xz")
