+++ requirements-lint.txt
@@ -3,4 +3,5 @@ flake8-bugbear==18.8.0
 flake8-docstrings==1.3.0
 flake8-import-order==0.18
 flake8-quotes==1.0.0
+flake8-per-file-ignores==0.6
 pylint==2.1.1
